// src/app/modules/room/janus/janus.interfaces.d.ts
declare namespace JanusJS {
// ...
  interface JSEP {}
// ...
  interface Publisher {
    id: number;
    display: string;
    streams: any[];
    audio_codec?: string;
    video_codec?: string;
    talking?: boolean;
  }

  enum MessageType {
    Recording = 'recording',
    Starting = 'starting',
    Started = 'started',
    Stopped = 'stopped',
    SlowLink = 'slow_link',
    Preparing = 'preparing',
    Refreshing = 'refreshing'
  }

  interface Message {
    result?: {
      status: MessageType;
      id?: string;
      uplink?: number;
    };
    error?: Error;
    publishers?: Publisher[];
  }

  interface PluginOptions {
    plugin: string;
    opaqueId?: string;
    dataChannelOptions?: RTCDataChannelInit;
    success?: (handle: PluginHandle) => void;
    error?: (error: any) => void;
    consentDialog?: (on: boolean) => void;
    webrtcState?: (isConnected: boolean) => void;
    iceState?: (state: 'connected' | 'failed' | 'disconnected' | 'closed') => void;
    mediaState?: (medium: 'audio' | 'video', receiving: boolean, mid?: number) => void;
    slowLink?: (uplink: boolean, lost: number) => void;
    onmessage?: (message: Message, jsep?: JSEP) => void;
    onlocaltrack?: (track: MediaStreamTrack, on: boolean) => void;
    onremotetrack?: (track: MediaStreamTrack, mid, on: boolean) => void;
    ondataopen?: Function;
    ondata?: Function;
    oncleanup?: Function;
    detached?: Function;
  }
// ...
  interface PluginMessage {
    message: {
      request: string;
      [otherProps: string]: any;
    };
    jsep?: JSEP;
    success?: Function;
    error?: (error: any) => void;
  }

  interface PluginHandle {
    plugin: string;
    id: string;
    token?: string;
    detached : boolean;
    webrtcStuff: {
      started: boolean,
      myStream: MediaStream,
      streamExternal: boolean,
      remoteStream: MediaStream,
      mySdp: any,
      mediaConstraints: any,
      pc: RTCPeerConnection,
      dataChannelOptions: RTCDataChannelInit,
      dataChannel: Array<RTCDataChannel>,
      dtmfSender: any,
      trickle: boolean,
      iceDone: boolean,
      volume: {
        value: number,
        timer: number
      }
    };
    getId(): string;
    getPlugin(): string;
    send(message: PluginMessage): void;
    createOffer(params: any): void;
    createAnswer(params: any): void;
    handleRemoteJsep(params: { jsep: JSEP }): void;
    dtmf(params: any): void;
    data(params: any): void;
    isAudioMuted(): boolean;
    muteAudio(): void;
    unmuteAudio(): void;
    isVideoMuted(): boolean;
    muteVideo(): void;
    unmuteVideo(): void;
    getBitrate(): string;
    hangup(sendRequest?: boolean): void;
    detach(params: any): void;
  }
// ...
}